Deepti_Singh/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
anthropic>=0.40.0
tavily-python>=0.5.0
duckduckgo-search>=6.0.0
arxiv>=2.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0
matplotlib>=3.9.0
plotly>=5.22.0
pandas>=2.0.0
numpy>=1.26.0
reportlab>=4.0.0
weasyprint>=62.0
markdown>=3.6.0
fastapi>=0.115.0
uvicorn>=0.30.0
sse-starlette>=2.0.0
aiofiles>=24.0.0
tenacity>=9.0.0
loguru>=0.7.0
rich>=13.0.0
mcp>=1.0.0
fastmcp>=3.2.0
googlesearch-python>=1.3.0
llama-index-core>=0.14.0
llama-index-readers-web>=0.6.0
playwright>=1.40.0
langchain-mcp-adapters>=0.2.0